+++ client/ayon_usd/standalone/usd/pinning/__init__.py
@@ -1,3 +1,3 @@
-from .pinning_file_helper import generate_pinning_file
+from .__pinning_file_generation_funcs import generate_pinning_file
 
 __all__ = ["generate_pinning_file"]
